support symfony 3.4, new testing app structure
