+++ .travis.yml
@@ -28,29 +28,28 @@ cache:
 env:
   matrix: SYMFONY_VERSION=^3.4@dev
   global:
-    - SYMFONY_DEPRECATIONS_HELPER=24
+    - SYMFONY_DEPRECATIONS_HELPER=0
     - SYMFONY_PHPUNIT_DIR=.phpunit SYMFONY_PHPUNIT_REMOVE="symfony/yaml"
+    - KERNEL_CLASS=Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Fixtures\App\Kernel
     - SYMFONY_PHPUNIT_VERSION=5.7
     - TEST_INSTALLATION=false
 
 matrix:
   include:
     - php: 7.1
+      env: DEPS=dev SYMFONY_VERSION=^4@dev
+    - php: 7.1
       env: DEPS=dev SYMFONY_VERSION=^3.4@dev
     - php: 7.1
-      env: DEPS=dev SYMFONY_VERSION=3.3.*
-    - php: 5.6
       env: COMPOSER_FLAGS="--prefer-lowest" SYMFONY_VERSION=2.8.* SYMFONY_DEPRECATIONS_HELPER=weak
-    - php: 7.0
-      env: SYMFONY_VERSION=3.1.*
-    - php: 7.0
-      env: SYMFONY_VERSION=3.2.*
+    - php: 7.1
+      env: SYMFONY_VERSION=3.3.*
     - env: TEST_INSTALLATION=true
 
   fast_finish: true
   allow_failures:
     - php: 7.1
-      env: DEPS=dev SYMFONY_VERSION=^3.4@dev
+      env: DEPS=dev SYMFONY_VERSION=^4@dev
     - env: TEST_INSTALLATION=true
