+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+* **2017-11-15**: Removed php 5.6 and 7.0 support, removed Symfony 3.0.* and 3.1.* support
+
+
 2.0.0-RC3 (unreleased)
 ----------------------
