+++ README.md
@@ -20,8 +20,8 @@ This bundle automatically creates and manages routes for configured persisted do
 
 ## Requirements
 
-* PHP 5.6 / 7.0 / 7.1
-* Symfony 2.8 / 3.1 / 3.2 / 3.3
+* PHP 7.1
+* Symfony 2.8 / 3.3 / ^3.4@dev
 * See also the `require` section of [composer.json](composer.json)
 
 ## Documentation
