+++ composer.json
@@ -11,19 +11,19 @@
     ],
     "require": {
         "php": "^5.6|^7.0",
-        "symfony/framework-bundle": "^2.8|^3.0",
-        "symfony-cmf/routing-auto": "^2.0.0-RC5",
+        "symfony/framework-bundle": "^2.8|^3.3",
+        "symfony-cmf/routing-auto": "^2.0",
         "symfony-cmf/routing-bundle": "^1.2.0|^2.0",
         "aferrandini/urlizer": "1.0.*",
-        "phpdocumentor/reflection-docblock": "^3.1, !=3.2.0",
-        "symfony-cmf/testing": "^2.1@dev"
+        "phpdocumentor/reflection-docblock": "^3.1, !=3.2.0"
     },
     "require-dev": {
         "symfony/phpunit-bridge": "^3.2",
-        "symfony/yaml": "^2.8|^3.0",
+        "symfony/yaml": "^2.8|^3.3",
         "matthiasnoback/symfony-dependency-injection-test": "~0.6",
         "matthiasnoback/symfony-config-test": "^1.3.1",
-        "doctrine/phpcr-odm": "^1.3"
+        "symfony-cmf/testing": "^2.1@dev",
+        "doctrine/phpcr-odm": "^1.4|^2.0"
     },
     "suggest": {
         "doctrine/phpcr-odm": "To enable support for the PHPCR ODM documents",
@@ -50,5 +50,7 @@
             "vendor/symfony-cmf/testing/bin/travis/phpcr_odm_doctrine_dbal.sh",
             "SYMFONY_PHPUNIT_REMOVE=\"symfony/yaml\" vendor/bin/simple-phpunit"
         ]
-    }
+    },
+    "minimum-stability":"dev",
+    "prefer-stable": true
 }
