+++ phpunit.xml.dist
@@ -30,8 +30,4 @@
         </whitelist>
     </filter>
 
-    <php>
-        <server name="KERNEL_DIR" value="tests/Resources/app" />
-    </php>
-
 </phpunit>
