+++ src/DependencyInjection/Compiler/AdapterPass.php
@@ -59,5 +59,6 @@ class AdapterPass implements CompilerPassInterface
 
         $managerDef = $container->getDefinition('cmf_routing_auto.auto_route_manager');
         $container->setAlias('cmf_routing_auto.adapter', $adapterId);
+        $container->getAlias('cmf_routing_auto.adapter')->setPublic(true);
     }
 }
