+++ tests/Fixtures/App/Document/AbstractContent.php
@@ -9,16 +9,14 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document;
+namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Fixtures\App\Document;
 
 use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCR;
 
 /**
- * @PHPCR\Document(
- *      referenceable=true
- * )
+ * @PHPCR\Document(referenceable=true)
  */
-class Blog
+class AbstractContent
 {
     /**
      * @PHPCR\Id()
