+++ tests/Fixtures/App/Document/ConflictProneArticle.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document;
+namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Fixtures\App\Document;
 
 use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCR;
 
 /**
  * @PHPCR\Document(translator="child", referenceable=true)
  */
-class SeoArticleMultilang extends Article
+class ConflictProneArticle extends Article
 {
 }
