+++ tests/Fixtures/App/Document/Page.php
@@ -9,13 +9,15 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document;
+namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Fixtures\App\Document;
 
 use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCR;
 
 /**
- * @PHPCR\Document(translator="child", referenceable=true)
+ * Document mapped to /{title} schema.
+ *
+ * @PHPCR\Document(referenceable=true)
  */
-class ConflictProneArticle extends Article
+class Page extends SeoArticle
 {
 }
