+++ tests/Fixtures/fixtures/loader_config/invalid1.xml
