+++ tests/Fixtures/fixtures/loader_config/invalid2.yml
