+++ tests/Fixtures/fixtures/loader_config/invalid3.yml
