+++ tests/Fixtures/fixtures/loader_config/valid1.xml
