+++ tests/Fixtures/fixtures/loader_config/valid2.xml
