+++ tests/Fixtures/fixtures/loader_config/valid3.xml
