+++ tests/Fixtures/fixtures/loader_config/valid4.xml
