+++ tests/Fixtures/fixtures/loader_config/valid5.xml
