+++ tests/Functional/Command/RefreshCommandTest.php
@@ -12,9 +12,9 @@
 namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Functional\Command;
 
 use Symfony\Cmf\Bundle\RoutingAutoBundle\Command\RefreshCommand;
+use Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Fixtures\App\Document\Blog;
+use Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Fixtures\App\Document\Post;
 use Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Functional\BaseTestCase;
-use Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document\Blog;
-use Symfony\Cmf\Bundle\RoutingAutoBundle\Tests\Resources\Document\Post;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Output\NullOutput;
 use Symfony\Component\Console\Output\StreamOutput;
