+++ tests/Unit/DependencyInjection/ConfigurationTest.php
@@ -47,7 +47,7 @@ class ConfigurationTest extends AbstractExtensionConfigurationTestCase
         ];
 
         $sources = array_map(function ($path) {
-            return __DIR__.'/../../Resources/Fixtures/'.$path;
+            return __DIR__.'/../../Fixtures/fixtures/'.$path;
         }, [
             'config/config.yml',
             'config/config.xml',
